% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{AICc}
\alias{AICc}
\title{Function to return corrected AIC from a fitted object}
\usage{
AICc(object)
}
\arguments{
\item{object}{model object}
}
\value{
AICc
}
\description{
Function to return corrected AIC from a fitted object
}
