% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AnnualFrequencyNatPlantNZ}
\alias{AnnualFrequencyNatPlantNZ}
\title{Annual frequency count of NZ (north island) naturalised plants}
\format{A data frame with columns:
\describe{
 \item{Year}{Year of records.}
 \item{No_speciemen}{Number of species recorded per year.}
}}
\source{
Aikio et al. \url{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1600-0706.2009.17963.x}
}
\usage{
AnnualFrequencyNatPlantNZ
}
\description{
Annual frequency count of NZ (north island) naturalised plants
}
\examples{
\dontrun{
 AnnualFrequencyNatPlantNZ
}
}
\keyword{datasets}
