% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{freqplot}
\alias{freqplot}
\title{Function to plot the frequency}
\usage{
freqplot(fit1, fit2 = NULL, fit3 = NULL, fit4 = NULL,
  xlab = "Year", ylab = "Frequency", main = fit1$name, cols = 2:5,
  ...)
}
\arguments{
\item{fit1}{the model fit1}

\item{fit2}{the model fit2}

\item{fit3}{the model fit3}

\item{fit4}{the model fit4}

\item{xlab}{x lable}

\item{ylab}{y lable}

\item{main}{the title}

\item{cols}{numberof columns}

\item{...}{further arguments to be passed to the predict funtion}
}
\value{
plot
}
\description{
Function to plot the frequency
}
