% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GBIFraw}
\alias{GBIFraw}
\title{GBIF raw data obtained via rgbif}
\format{A data frame with columns
\describe{
 \item{.id}{ID}
 \item{name}{name}
 \item{key}{key}
 \item{basisOfRecord}{basisOfRecord}
 \item{year}{year}
 \item{month}{month}
 \item{day}{day}
 \item{issues}{issues}
 \item{countryCode}{countryCode}
 \item{gbifID}{gbifID}
 \item{verbatimLocality}{verbatimLocality}
 \item{occurrenceID}{occurrenceID}
 \item{coordinateUncertaintyInMeters}{coordinateUncertaintyInMeters}
 \item{continent}{continent}
 \item{stateProvince}{stateProvince}
 \item{coordinatePrecision}{coordinatePrecision}
}}
\source{
gbif.org \url{https://gbif.org}
}
\usage{
GBIFraw
}
\description{
GBIF raw data obtained via rgbif
}
\examples{
\dontrun{
 GBIFraw
}
}
\keyword{datasets}
