% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifwrangler.R
\name{gbifwrangler}
\alias{gbifwrangler}
\title{Prepare GBIF data for lag test}
\usage{
gbifwrangler(x, year = 2017, minocc = 15, noyears = 5)
}
\arguments{
\item{x}{GBIF data.frame obtained using rgibf package}

\item{year}{cut off year to truncate dataset to cater for the delay between surveys and submission to GBIF. Defaults to 2017.}

\item{minocc}{minimum occurences per species (defaults to 15)}

\item{noyears}{number of years with occurences (defauls to 5). Set to 0 if you do not want to set a minumum.}
}
\value{
timeseries data.frame for use as x in lag test

frequency data.frame for use as y in lag test
}
\description{
Prepare GBIF data for lag test
}
\examples{
{
 gbifwrangler(GBIFraw, year=2017, minocc = 15, noyears=5)
}


}
