% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.species}
\alias{get.species}
\title{get species from time series objects}
\usage{
get.species(x, y, species, zeros = FALSE)
}
\arguments{
\item{x}{data.frame with cumulative record counts per year}

\item{y}{data.frame with occurence records per species and year}

\item{species}{#vector of species or other time series objects}

\item{zeros}{binomimal if to treat missing data as 0. If zeros=TRUE, include zeros in returned data.}
}
\value{
results data.frame with species
}
\description{
get species from time series objects
}
