% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lagphase.knots}
\alias{lagphase.knots}
\title{Function model with lag phase followed by growth where knots and order are specified}
\usage{
lagphase.knots(knots, data, order)
}
\arguments{
\item{knots}{the numbr of nots}

\item{data}{data.frame}

\item{order}{the order of the knots}
}
\value{
results data.frame with species
}
\description{
Function model with lag phase followed by growth where knots and order are specified
}
