% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lagphase}
\alias{lagphase}
\title{Main function fitting glms to detect get end of constant population growth.
Give it a set of data where the columns include year, frequency, specimens
It will find appropriate knots if not specified. It will choose an appropriate
order if not specified. Just don't give it knots but no order
If gam=TRUE, it will return a gam model instead.}
\usage{
lagphase(data, knots = NULL, order = 1, gam = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{knots}{the numbr of nots}

\item{order}{the order of the knots}

\item{gam}{If gam=TRUE, it will return a gam model instead}
}
\value{
results data.frame with species
}
\description{
Main function fitting glms to detect get end of constant population growth.
Give it a set of data where the columns include year, frequency, specimens
It will find appropriate knots if not specified. It will choose an appropriate
order if not specified. Just don't give it knots but no order
If gam=TRUE, it will return a gam model instead.
}
