% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lagphaseplot}
\alias{lagphaseplot}
\title{Function to plot the fitted spline function after adjusting for number of specimens (records)}
\usage{
lagphaseplot(fit, ylim = NULL, xlab = "Year",
  ylab = "Adjusted frequency", main = fit$name, ...)
}
\arguments{
\item{fit}{the model fit}

\item{ylim}{limit of y axis}

\item{xlab}{x lable}

\item{ylab}{y lable}

\item{main}{the title}

\item{...}{further arguments to be passed to the predict funtion}
}
\value{
plot
}
\description{
Function to plot the fitted spline function after adjusting for number of specimens (records)
}
