% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pwlm}
\alias{pwlm}
\title{Fit piecewise linear model}
\usage{
pwlm(x, y)
}
\arguments{
\item{x}{time series object}

\item{y}{frequency counts}
}
\value{
results
}
\description{
Fit piecewise linear model
}
