% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runlag.R
\name{runlag}
\alias{runlag}
\title{Time lag detection}
\usage{
runlag(x, y)
}
\arguments{
\item{x}{data.frame with cumulative record counts per year}

\item{y}{data.frame with occurence records per species and year}
}
\value{
results data.frame with lag assessment results
}
\description{
Time lag detection
}
\examples{
\dontrun{
 runlag(x,y)
}
}
