% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TimeSeriesNatPlantNZN}
\alias{TimeSeriesNatPlantNZN}
\title{Time series data of NZN (north island) naturalised plants}
\format{A data frame with columns:
\describe{
 \item{Species}{String to separate species.}
 \item{Year}{Year of records.}
 \item{COUNT}{Number of species recorded per year. }
 \item{CUM_FREQ}{Cumulative frequency: total count of species per year.}
}}
\source{
Aikio et al. \url{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1600-0706.2009.17963.x}
}
\usage{
TimeSeriesNatPlantNZN
}
\description{
Time series data of NZN (north island) naturalised plants
}
\examples{
\dontrun{
 TimeSeriesNatPlantNZN
}
}
\keyword{datasets}
