% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tryknots}
\alias{tryknots}
\title{Function that checks that the specified knots make sense then use lagphase.knots
to fit the model and returns AIC of fitted model}
\usage{
tryknots(knots, data, order)
}
\arguments{
\item{knots}{the numbr of nots}

\item{data}{data.frame}

\item{order}{the order of the knots}
}
\value{
AIC of fitted model
}
\description{
Function that checks that the specified knots make sense then use lagphase.knots
to fit the model and returns AIC of fitted model
}
