% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GBIFraw}
\alias{GBIFraw}
\title{GBIF raw data obtained via rgbif}
\format{A data frame with columns
\describe{
 \item{Year}{Year of records.}
 \item{No_speciemen}{Number of species recorded per year.}
}}
\source{
gbif.org \url{https://gbif.org}
}
\usage{
GBIFraw
}
\description{
GBIF raw data obtained via rgbif
}
\examples{
\dontrun{
 GBIFraw
}
}
\keyword{datasets}
